requirements.txt
grpcio
requests
numpy
google
protobuf
toml
watchdog
pandas
mappy
pyguppy
